@@ -427,20 +427,24 @@ class ExecuteTest(fixtures.TablesTest):
         # TODO: this test is assuming too much of arbitrary dialects and would
         # be better suited tested against a single mock dialect that does not
         # have any special behaviors
-        with patch.object(
-            testing.db.dialect, "dbapi", Mock(Error=DBAPIError)
-        ), patch.object(
-            testing.db.dialect, "loaded_dbapi", Mock(Error=DBAPIError)
-        ), patch.object(
-            testing.db.dialect, "is_disconnect", lambda *arg: False
-        ), patch.object(
-            testing.db.dialect,
-            "do_execute",
-            Mock(side_effect=NonStandardException),
-        ), patch.object(
-            testing.db.dialect.execution_ctx_cls,
-            "handle_dbapi_exception",
-            Mock(),
+        with (
+            patch.object(testing.db.dialect, "dbapi", Mock(Error=DBAPIError)),
+            patch.object(
+                testing.db.dialect, "loaded_dbapi", Mock(Error=DBAPIError)
+            ),
+            patch.object(
+                testing.db.dialect, "is_disconnect", lambda *arg: False
+            ),
+            patch.object(
+                testing.db.dialect,
+                "do_execute",
+                Mock(side_effect=NonStandardException),
+            ),
+            patch.object(
+                testing.db.dialect.execution_ctx_cls,
+                "handle_dbapi_exception",
+                Mock(),
+            ),
         ):
             with testing.db.connect() as conn:
                 assert_raises(
@@ -1001,11 +1005,14 @@ class ConvenienceExecuteTest(fixtures.TablesTest):
         engine = engines.testing_engine()
 
         close_mock = Mock()
-        with mock.patch.object(
-            engine._connection_cls,
-            "begin",
-            Mock(side_effect=Exception("boom")),
-        ), mock.patch.object(engine._connection_cls, "close", close_mock):
+        with (
+            mock.patch.object(
+                engine._connection_cls,
+                "begin",
+                Mock(side_effect=Exception("boom")),
+            ),
+            mock.patch.object(engine._connection_cls, "close", close_mock),
+        ):
             with expect_raises_message(Exception, "boom"):
                 with engine.begin():
                     pass
@@ -1894,11 +1901,12 @@ class EngineEventsTest(fixtures.TestBase):
             # as part of create
             # note we can't use an event to ensure begin() is not called
             # because create also blocks events from happening
-            with mock.patch.object(
-                e1.dialect, "initialize", side_effect=init
-            ) as m1, mock.patch.object(
-                e1._connection_cls, "begin"
-            ) as begin_mock:
+            with (
+                mock.patch.object(
+                    e1.dialect, "initialize", side_effect=init
+                ) as m1,
+                mock.patch.object(e1._connection_cls, "begin") as begin_mock,
+            ):
 
                 @event.listens_for(e1, "connect", insert=True)
                 def go1(dbapi_conn, xyz):
@@ -2536,11 +2544,14 @@ class EngineEventsTest(fixtures.TestBase):
         def conn_tracker(conn, opt):
             opt["conn_tracked"] = True
 
-        with mock.patch.object(
-            engine.dialect, "set_connection_execution_options"
-        ) as conn_opt, mock.patch.object(
-            engine.dialect, "set_engine_execution_options"
-        ) as engine_opt:
+        with (
+            mock.patch.object(
+                engine.dialect, "set_connection_execution_options"
+            ) as conn_opt,
+            mock.patch.object(
+                engine.dialect, "set_engine_execution_options"
+            ) as engine_opt,
+        ):
             e2 = engine.execution_options(e1="opt_e1")
             c1 = engine.connect()
             c2 = c1.execution_options(c1="opt_c1")
@@ -3493,11 +3504,12 @@ class OnConnectTest(fixtures.TestBase):
             nonlocal init_connection
             init_connection = connection
 
-        with mock.patch.object(
-            e._connection_cls, "begin"
-        ) as mock_begin, mock.patch.object(
-            e.dialect, "initialize", Mock(side_effect=mock_initialize)
-        ) as mock_init:
+        with (
+            mock.patch.object(e._connection_cls, "begin") as mock_begin,
+            mock.patch.object(
+                e.dialect, "initialize", Mock(side_effect=mock_initialize)
+            ) as mock_init,
+        ):
             conn = e.connect()
 
             eq_(mock_begin.mock_calls, [])
@@ -3928,12 +3940,16 @@ class SetInputSizesTest(fixtures.TablesTest):
         # "safe" datatypes so that the DBAPI does not actually need
         # setinputsizes() called in order to work.
 
-        with mock.patch.object(
-            engine.dialect, "bind_typing", BindTyping.SETINPUTSIZES
-        ), mock.patch.object(
-            engine.dialect, "do_set_input_sizes", do_set_input_sizes
-        ), mock.patch.object(
-            engine.dialect.execution_ctx_cls, "pre_exec", pre_exec
+        with (
+            mock.patch.object(
+                engine.dialect, "bind_typing", BindTyping.SETINPUTSIZES
+            ),
+            mock.patch.object(
+                engine.dialect, "do_set_input_sizes", do_set_input_sizes
+            ),
+            mock.patch.object(
+                engine.dialect.execution_ctx_cls, "pre_exec", pre_exec
+            ),
         ):
             yield engine, canary
 
