@@ -1577,11 +1577,14 @@ class DeclarativeMultiBaseTest(
             attr_type.fail()
 
     def test_column_named_twice(self):
-        with expect_warnings(
-            "On class 'Foo', Column object 'x' named directly multiple "
-            "times, only one will be used: x, y. Consider using "
-            "orm.synonym instead"
-        ), expect_raises(exc.DuplicateColumnError):
+        with (
+            expect_warnings(
+                "On class 'Foo', Column object 'x' named directly multiple "
+                "times, only one will be used: x, y. Consider using "
+                "orm.synonym instead"
+            ),
+            expect_raises(exc.DuplicateColumnError),
+        ):
 
             class Foo(Base):
                 __tablename__ = "foo"
@@ -1592,11 +1595,14 @@ class DeclarativeMultiBaseTest(
 
     @testing.variation("style", ["old", "new"])
     def test_column_repeated_under_prop(self, style):
-        with expect_warnings(
-            "On class 'Foo', Column object 'x' named directly multiple "
-            "times, only one will be used: x, y, z. Consider using "
-            "orm.synonym instead"
-        ), expect_raises(exc.DuplicateColumnError):
+        with (
+            expect_warnings(
+                "On class 'Foo', Column object 'x' named directly multiple "
+                "times, only one will be used: x, y, z. Consider using "
+                "orm.synonym instead"
+            ),
+            expect_raises(exc.DuplicateColumnError),
+        ):
             if style.old:
 
                 class Foo(Base):
