@@ -2712,8 +2712,9 @@ class MultipleAdaptUsesEntityOverTableTest(
     def test_two_joins_adaption(self):
         a, c, d = self.tables.a, self.tables.c, self.tables.d
 
-        with _aliased_join_warning(r"C\(c\)"), _aliased_join_warning(
-            r"D\(d\)"
+        with (
+            _aliased_join_warning(r"C\(c\)"),
+            _aliased_join_warning(r"D\(d\)"),
         ):
             q = self._two_join_fixture()._compile_state()
 
@@ -2745,8 +2746,9 @@ class MultipleAdaptUsesEntityOverTableTest(
     def test_two_joins_sql(self):
         q = self._two_join_fixture()
 
-        with _aliased_join_warning(r"C\(c\)"), _aliased_join_warning(
-            r"D\(d\)"
+        with (
+            _aliased_join_warning(r"C\(c\)"),
+            _aliased_join_warning(r"D\(d\)"),
         ):
             self.assert_compile(
                 q,
