@@ -429,9 +429,12 @@ class VersioningTest(fixtures.MappedTest):
             else:
                 return self.context.rowcount
 
-        with patch.object(
-            config.db.dialect, "supports_sane_multi_rowcount", False
-        ), patch("sqlalchemy.engine.cursor.CursorResult.rowcount", rowcount):
+        with (
+            patch.object(
+                config.db.dialect, "supports_sane_multi_rowcount", False
+            ),
+            patch("sqlalchemy.engine.cursor.CursorResult.rowcount", rowcount),
+        ):
             Foo = self.classes.Foo
             s1 = self._fixture()
             f1s1 = Foo(value="f1 value")
@@ -444,10 +447,11 @@ class VersioningTest(fixtures.MappedTest):
             eq_(f1s1.version_id, 2)
 
     def test_update_delete_no_plain_rowcount(self):
-        with patch.object(
-            config.db.dialect, "supports_sane_rowcount", False
-        ), patch.object(
-            config.db.dialect, "supports_sane_multi_rowcount", False
+        with (
+            patch.object(config.db.dialect, "supports_sane_rowcount", False),
+            patch.object(
+                config.db.dialect, "supports_sane_multi_rowcount", False
+            ),
         ):
             Foo = self.classes.Foo
             s1 = self._fixture()
@@ -714,10 +718,11 @@ class VersionOnPostUpdateTest(fixtures.MappedTest):
 
         n1.related.append(n2)
 
-        with patch.object(
-            config.db.dialect, "supports_sane_rowcount", False
-        ), patch.object(
-            config.db.dialect, "supports_sane_multi_rowcount", False
+        with (
+            patch.object(config.db.dialect, "supports_sane_rowcount", False),
+            patch.object(
+                config.db.dialect, "supports_sane_multi_rowcount", False
+            ),
         ):
             s2 = Session(bind=s.connection(bind_arguments=dict(mapper=Node)))
             s2.query(Node).filter(Node.id == n2.id).update({"version_id": 3})
